Make Codensity kind and levity polymorphic
With `TypeInType`, we can get

```haskell
Codensity :: (k -> TYPE rep) -> Type -> Type
```

I don't know if the polymorphism in `k` is useful; it was
readily available, so I took it. I actually have an application in
mind for polymorphism in `rep`. The `primitive` package offers

```haskell
indexArrayM :: Monad m => Array a -> Int -> m a
```

Now what if I want to recover the underlying functionality of
`indexArray#` from that? One option would be to use the
`data Box a = Box a` "lifted identity" monad. Another would be
to use a custom-written `Cont` monad (a clone of the one from
`transformers`). But I realized that `Codensity` *almost* works
off the shelf, if we just add a bit of extra polymorphism.
With this patch, I can write

```haskell
iarr# :: Array a -> Int -> (# a #)
iarr# ar i = runCodensity (indexArrayM ar i) (\a -> (# a #))
```

and get exactly the Core I want

```
iarr# :: forall a. Array a -> Int -> (# a #)
[GblId,
 Arity=2,
 Caf=NoCafRefs,
 Str=<S(S),1*U(U)><S(S),1*U(U)>,
 Unf=Unf{Src=InlineStable, TopLvl=True, Value=True, ConLike=True,
         WorkFree=True, Expandable=True,
         Guidance=ALWAYS_IF(arity=2,unsat_ok=True,boring_ok=False)
         Tmpl= \ (@ a_aetV)
                 (ar_aetE [Occ=Once!] :: Array a_aetV)
                 (i_aetF [Occ=Once!] :: Int) ->
                 case i_aetF of { GHC.Types.I# i#_aeun [Occ=Once] ->
                 case ar_aetE of { Array ds2_aeur [Occ=Once] ->
                 GHC.Prim.indexArray# @ a_aetV ds2_aeur i#_aeun
                 }
                 }}]
iarr#
  = \ (@ a_aetV) (ar_aetE :: Array a_aetV) (i_aetF :: Int) ->
      case i_aetF of { GHC.Types.I# i#_aeun ->
      case ar_aetE of { Array ds2_aeur ->
      GHC.Prim.indexArray# @ a_aetV ds2_aeur i#_aeun
      }
      }
```